2. YT Automation/pyproject.toml
[tool.poetry]
name = "crewai-yt-automation"
version = "0.1.0"
description = "Automate YouTube tasks with CrewAI"
authors = ["Vasanth"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "^0.19.0"}
pydantic = "^2.6.3"
load-dotenv = "^0.1.0"
crewai-tools = "^0.0.15"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"